+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+* Add `aws-sso-cli completion --source` flag to generate completion script and
+  print to stdout. #779
+
 ## [v1.14.3] - 2024-01-15
 
 ### Bugs